rewrote file structure to put models in a seperate file and routes in a seperate file, can revert back if this is not useful
